requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
DateTime==4.2
googlefinance==0.7
googlefinance.client==1.3.0
idna==2.6
kafka-python==1.3.5
numpy==1.13.3
pandas==0.20.3
py4j==0.10.4
pyspark==2.2.0
python-dateutil==2.6.1
pytz==2017.2
realtime-stock==1.0.0
redis==2.10.6
requests==2.18.4
schedule==0.4.3
six==1.11.0
urllib3==1.22
zope.interface==4.4.3
